# docs/openapi.yaml
openapi: 3.1.0
info:
  title: LabelVerify API
  version: "0.1.0"
  description: |
    HTTP surface for the LabelVerify prototype. Two primary endpoints —
    `/api/verify` (compare a label image against declared application
    data) and `/api/extract` (read fields from a label image without a
    verdict). Auxiliary endpoints cover health, batch submission, and
    review queue inspection.

    All `/api/*` endpoints emit `X-Request-Id` on the response, echoing
    the inbound id if the client sent a valid one. Quote the id when
    reporting failures.
  contact:
    name: LabelVerify maintainer
    email: xandermlopez@gmail.com
servers:
  - url: https://label-verify-six.vercel.app
    description: Production demo
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    debugBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-string
      description: |
        Only required for `/api/debug/*` routes. Set via the
        `DEBUG_TOKEN` env var in the deployment.
  schemas:
    NetContents:
      type: object
      required: [value, unit]
      properties:
        value:
          type: number
          minimum: 0
        unit:
          type: string
          enum: [ml, L, fl_oz, gal, qt, pt]
    ProducerAddress:
      type: object
      properties:
        name: { type: [string, "null"] }
        street: { type: [string, "null"] }
        city: { type: [string, "null"] }
        state: { type: [string, "null"] }
        postal_code: { type: [string, "null"] }
        country: { type: [string, "null"] }
    DeclaredFields:
      type: object
      required:
        [brand_name, class_type, class_category, abv_percent, net_contents,
         producer, country_of_origin]
      properties:
        brand_name: { type: string, minLength: 1 }
        class_type: { type: string, minLength: 1 }
        class_category:
          type: string
          enum: [beer, wine, distilled_spirits, fortified_wine]
        abv_percent:
          type: number
          minimum: 0
          maximum: 100
        net_contents:
          $ref: "#/components/schemas/NetContents"
        producer:
          oneOf:
            - $ref: "#/components/schemas/ProducerAddress"
            - type: string
        country_of_origin:
          type: string
          minLength: 2
    FieldComparison:
      type: object
      required: [field, status, expected, actual, confidence]
      properties:
        field: { type: string }
        status:
          type: string
          enum: [pass, fail, review]
        expected: {}
        actual: {}
        confidence:
          type: number
          minimum: 0
          maximum: 1
        reason: { type: string }
        components:
          type: object
          additionalProperties:
            type: string
            enum: [pass, fail, review]
    GovernmentWarning:
      type: object
      required: [status, subscores]
      properties:
        status:
          type: string
          enum: [pass, fail, review]
        subscores:
          type: object
          required: [text, caps, bold, size]
          properties:
            text:
              type: object
              properties:
                status: { type: string, enum: [pass, fail, review] }
                confidence: { type: number, minimum: 0, maximum: 1 }
            caps:
              type: object
              properties:
                status: { type: string, enum: [pass, fail, review] }
                confidence: { type: number, minimum: 0, maximum: 1 }
            bold:
              type: object
              properties:
                status: { type: string, enum: [pass, fail, review] }
                confidence: { type: number, minimum: 0, maximum: 1 }
            size:
              type: object
              properties:
                status: { type: string, enum: [pass, fail, review] }
                confidence: { type: number, minimum: 0, maximum: 1 }
        reason: { type: string }
    VerifyResponse:
      type: object
      required:
        [verdict, imageQuality, fields, governmentWarning, extracted,
         timings, modelId, modelVersion, requiresHumanReview, reviewReasons]
      properties:
        verdict:
          type: string
          enum: [pass, fail, review]
        imageQuality:
          type: string
          enum: [good, low, bad]
        imageQualityReason: { type: string }
        fields:
          type: object
          required:
            [brand_name, class_type, abv_percent, net_contents, producer,
             country_of_origin]
          properties:
            brand_name: { $ref: "#/components/schemas/FieldComparison" }
            class_type: { $ref: "#/components/schemas/FieldComparison" }
            abv_percent: { $ref: "#/components/schemas/FieldComparison" }
            net_contents: { $ref: "#/components/schemas/FieldComparison" }
            producer: { $ref: "#/components/schemas/FieldComparison" }
            country_of_origin: { $ref: "#/components/schemas/FieldComparison" }
        governmentWarning:
          $ref: "#/components/schemas/GovernmentWarning"
        extracted:
          type: object
          description: Raw extractor output (per-field with confidence).
        timings:
          type: object
          required: [preprocess, vision, matching, total]
          properties:
            preprocess: { type: number }
            ocr:
              type: [number, "null"]
            vision: { type: number }
            matching: { type: number }
            total: { type: number }
        modelId: { type: string }
        modelVersion: { type: string }
        modeUsed: { type: string }
        requiresHumanReview: { type: boolean }
        reviewReasons:
          type: array
          items: { type: string }
        fallbackUsed: { type: boolean }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        aborted: { type: boolean }
        requestId: { type: string }
paths:
  /api/verify:
    post:
      summary: Verify a label image against declared COLA application data
      description: |
        Accepts the image as multipart/form-data (`image` file +
        `declared` JSON string) OR as application/json with a remote
        `url` field. Returns a structured per-field PASS / FAIL /
        REVIEW verdict plus the Government Warning subscore breakdown.

        Rate-limited per IP (default 60/min). Sends
        `X-RateLimit-Remaining` on success and `Retry-After` on 429.
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image, declared]
              properties:
                image:
                  type: string
                  format: binary
                declared:
                  type: string
                  description: JSON-encoded DeclaredFields
                mode:
                  type: string
                  description: Optional model-mode id (e.g. "default").
          application/json:
            schema:
              type: object
              required: [url, declared]
              properties:
                url: { type: string, format: uri }
                declared: { $ref: "#/components/schemas/DeclaredFields" }
                mode: { type: string }
      responses:
        "200":
          description: Verification succeeded
          headers:
            X-Request-Id:
              schema: { type: string }
            X-RateLimit-Remaining:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VerifyResponse" }
        "400":
          description: Malformed request (missing field, bad JSON, schema fail)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "413":
          description: Image / PDF too large
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "415":
          description: Unsupported MIME type
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: Rate-limited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          description: Verification failed (unexpected error)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: URL fetch failed (SSRF guard, remote 4xx/5xx)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "504":
          description: Vision call exceeded budget
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...